refactor: rename name_conflicts.py for clarity
